pyproject.toml
[project]
name = "azure-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.12.14",
    "aiosqlite==0.21.0",
    "azure-ai-agents==1.0.0",
    "azure-ai-projects==1.0.0b12",
    "azure-identity==1.23.0",
    "httpx==0.28.1",
    "ipykernel==6.29.5",
    "mcp==1.11.0",
    "openai==1.75.0",
    "openai-agents==0.0.11",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "uvicorn==0.35.0",
]
